Revert binary_sensor in nest api
